requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
python-multipart
pillow
websockets
aiohttp
aiofiles
httpx
litellm

# Testing
pytest
pytest-asyncio
pytest-timeout

